+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/action/SearchJqlWildcardAction.kt
@@ -1,5 +1,6 @@
 package com.atlassian.performance.tools.jiraactions.api.action
 
+import com.atlassian.performance.tools.jiraactions.api.SEARCH_WITH_JQL_WILDCARD
 import com.atlassian.performance.tools.jiraactions.api.WebJira
 import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 import com.atlassian.performance.tools.jiraactions.api.memories.IssueKeyMemory
@@ -7,21 +8,20 @@ import com.atlassian.performance.tools.jiraactions.api.memories.JqlMemory
 import com.atlassian.performance.tools.jiraactions.api.observation.SearchJqlObservation
 import com.atlassian.performance.tools.jiraactions.api.page.IssueNavigatorPage
 import com.atlassian.performance.tools.jiraactions.jql.BuiltInJQL
-import com.atlassian.performance.tools.jiraactions.api.SEARCH_WITH_JQL_PREV_REPORTER
 import java.util.function.Predicate
 import javax.json.JsonObject
 
-class SearchJqlPreviousReporter(
+class SearchJqlWildcardAction(
     private val jira: WebJira,
     private val meter: ActionMeter,
     private val jqlMemory: JqlMemory,
     private val issueKeyMemory: IssueKeyMemory
 ) : Action {
-
     override fun run() {
-        val jqlQuery = jqlMemory.recallByTag(Predicate { it == BuiltInJQL.REPORTERS.name })!!
+        val jqlQuery = jqlMemory.recallByTag(Predicate { it == BuiltInJQL.GENERIC_WIDE.name })!!
+
         val issueNavigatorPage = meter.measure(
-            key = SEARCH_WITH_JQL_PREV_REPORTER,
+            key = SEARCH_WITH_JQL_WILDCARD,
             action = { jira.goToIssueNavigator(jqlQuery).waitForIssueNavigator() },
             observation = this::observe
         )